Fix wrong service data QByteArray sub-array handling
Sub array is currently taken with +[number], as if it was a standard
C++ array. But in our case, it's QByteArray. This leads to a cast to
char* and that leads to zero-terminated data when converting again
back to QByteArray.

Pick-to: 6.5
Fixes: QTBUG-124130
Change-Id: I0a1850dd36d2158c7a5807a72b021b6875c009b8
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit 7718d0faa3ce590ad20417c94ff407ec5e0cc453)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>